removed short from whole project

+++ Client/game/blocks/clientBlocks.cpp
@@ -145,8 +145,8 @@ void ClientBlocks::render() {
     }
     
     int block_index = 0;
-    for(unsigned short x = getViewBeginX(); x < getViewEndX(); x++)
-        for(unsigned short y = getViewBeginY(); y < getViewEndY(); y++) {
+    for(int x = getViewBeginX(); x < getViewEndX(); x++)
+        for(int y = getViewBeginY(); y < getViewEndY(); y++) {
             if(getState(x, y) == 16)
                 updateState(x, y);
             
@@ -155,8 +155,8 @@ void ClientBlocks::render() {
                 int texture_x = (getRenderBlock(x, y)->variation) % (resource_pack->getTextureRectangle(getBlockType(x, y)).w / BLOCK_WIDTH) * BLOCK_WIDTH;
                 int texture_y = resource_pack->getTextureRectangle(getBlockType(x, y)).y + BLOCK_WIDTH * getRenderBlock(x, y)->state;
                 
-                block_rects.setTextureCoords(block_index, {(short)texture_x, (short)texture_y, BLOCK_WIDTH, BLOCK_WIDTH});
-                block_rects.setRect(block_index, {(short)block_x, (short)block_y, BLOCK_WIDTH * 2, BLOCK_WIDTH * 2});
+                block_rects.setTextureCoords(block_index, {texture_x, texture_y, BLOCK_WIDTH, BLOCK_WIDTH});
+                block_rects.setRect(block_index, {block_x, block_y, BLOCK_WIDTH * 2, BLOCK_WIDTH * 2});
                 
                 block_index++;
             }
@@ -165,11 +165,11 @@ void ClientBlocks::render() {
     if(block_index)
         block_rects.render(block_index, &resource_pack->getBlockTexture());
     
-    for(unsigned short x = getViewBeginX(); x < getViewEndX(); x++)
-        for(unsigned short y = getViewBeginY(); y < getViewEndY(); y++) {
+    for(int x = getViewBeginX(); x < getViewEndX(); x++)
+        for(int y = getViewBeginY(); y < getViewEndY(); y++) {
             if(getBreakStage(x, y)) {
                 int block_x = x * BLOCK_WIDTH * 2 - view_x + gfx::getWindowWidth() / 2, block_y = y * BLOCK_WIDTH * 2 - view_y + gfx::getWindowHeight() / 2;
-                resource_pack->getBreakingTexture().render(2, block_x, block_y, gfx::RectShape(0, short(BLOCK_WIDTH * (getBreakStage(x, y) - 1)), BLOCK_WIDTH, BLOCK_WIDTH));
+                resource_pack->getBreakingTexture().render(2, block_x, block_y, gfx::RectShape(0, BLOCK_WIDTH * (getBreakStage(x, y) - 1), BLOCK_WIDTH, BLOCK_WIDTH));
             }
         }
 }
