+++ Client/game/networking/clientNetworking.hpp
@@ -24,7 +24,7 @@ class ClientNetworking : public ClientModule, EventListener<ClientPacketEvent> {
     sf::TcpSocket socket;
     
     std::string ip_address, username;
-    unsigned short port;
+    int port;
     
     void onEvent(ClientPacketEvent& event) override;
     
@@ -33,7 +33,7 @@ class ClientNetworking : public ClientModule, EventListener<ClientPacketEvent> {
     void stop() override;
     void update(float frame_length) override;
 public:
-    ClientNetworking(const std::string& ip_address, unsigned short port, const std::string& username) : ip_address(ip_address), port(port), username(username) {}
+    ClientNetworking(const std::string& ip_address, int port, const std::string& username) : ip_address(ip_address), port(port), username(username) {}
     
     void sendPacket(sf::Packet& packet);
     std::vector<char> getData();
