+++ Client/menus/pauseScreen/pauseScreen.hpp
@@ -24,6 +24,6 @@ public:
     bool hasExitedToMenu() const;
     
     void renderBack() override;
-    void setBackWidth(unsigned short width) override;
-    unsigned short getBackWidth() override;
+    void setBackWidth(int width) override;
+    int getBackWidth() override;
 };
