+++ Client/menus/worldSelector/worldSelector.hpp
@@ -8,7 +8,7 @@ struct WorldToSelect {
     std::string name;
     gfx::Button play_button, delete_button;
     gfx::Texture title, icon, last_played;
-    void render(int position, unsigned short mouse_x, unsigned short mouse_y);
+    void render(int position, int mouse_x, int mouse_y);
     int y;
 };
