+++ Server/entities/serverEntities.cpp
@@ -15,7 +15,7 @@ void ServerEntities::stop() {
 
 void ServerEntities::onEvent(EntityDeletionEvent& event) {
     sf::Packet packet;
-    packet << ServerPacketType::ENTITY_DELETION << (unsigned short)event.entity->id;
+    packet << ServerPacketType::ENTITY_DELETION << event.entity->id;
     networking->sendToEveryone(packet);
 }
